html: split large module into smaller ones
